separate Tahoma and time limit login
